fix: address milestone 1 architecture review findings
- Fix server response header aliasing by cloning slices instead of
  assigning tape header slices directly to the response writer
- Make NewServer panic on nil store, consistent with NewRecorder
  (both are programming errors per CLAUDE.md "no panics" rule)
- Move Matcher, MatcherFunc, and ExactMatcher from server.go to
  matcher.go for discoverability alongside other matching logic
- Add doc.go with package-level documentation
- Update stale ADR-3 references in decisions.md to reflect ADR-4
  changes (DefaultMatcher as default, ExactMatcher superseded)
